# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "健康数据平台 API 文档",
        "title": "HealthDT API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/app/wechat_login": {
            "post": {
                "description": "微信授权码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "微信登录",
                "parameters": [
                    {
                        "description": "微信授权码",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.WechatLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/events": {
            "get": {
                "description": "获取所有事件信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "查询事件列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/health_data": {
            "post": {
                "description": "新增一条健康数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health_data"
                ],
                "summary": "创建健康数据记录",
                "parameters": [
                    {
                        "description": "健康数据内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HealthDataRecord"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health_data/{id}": {
            "get": {
                "description": "根据ID获取健康数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health_data"
                ],
                "summary": "查询健康数据记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "健康数据ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "根据ID更新健康数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health_data"
                ],
                "summary": "更新健康数据记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "健康数据ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "健康数据内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HealthDataRecord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除健康数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health_data"
                ],
                "summary": "删除健康数据记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "健康数据ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "根据ID查询用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取单个用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AppUser"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "description": "查询所有用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AppUser"
                            }
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "新用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AppUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AppUser"
                        }
                    },
                    "500": {
                        "description": "注册失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/update": {
            "put": {
                "description": "更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户信息更新",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AppUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AppUser"
                        }
                    },
                    "500": {
                        "description": "更新失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.WechatLoginRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "models.AppUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "is_active": {
                    "description": "激活状态",
                    "type": "boolean"
                },
                "last_login": {
                    "description": "最后登录时间",
                    "type": "string"
                },
                "password_hash": {
                    "description": "密码哈希",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                },
                "wechat_openid": {
                    "description": "微信openid",
                    "type": "string"
                }
            }
        },
        "models.DeviceAssignment": {
            "type": "object",
            "properties": {
                "assigned_at": {
                    "type": "string"
                },
                "device_id": {
                    "type": "integer"
                },
                "health_profile_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "unassigned_at": {
                    "type": "string"
                }
            }
        },
        "models.HealthDataRecord": {
            "type": "object"
        },
        "models.HealthProfile": {
            "type": "object"
        }
    }
}